Moved OSVRCustomPresent implementations into their own files. Previously these were included in OSVRHMD.h
