+++ OSVRUnreal/Plugins/OSVR/Source/OSVR/Private/OSVRCustomPresent.h
@@ -1,0 +1,106 @@
+//
+// Copyright 2016 Sensics, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+
+#pragma once
+
+#include "IOSVR.h"
+#include <osvr/RenderKit/RenderManagerC.h>
+#include <vector>
+
+DECLARE_LOG_CATEGORY_EXTERN(FOSVRCustomPresentLog, Log, All);
+
+template<class TGraphicsDevice>
+class FOSVRCustomPresent : public FRHICustomPresent
+{
+public:
+    FTexture2DRHIRef mRenderTexture;
+
+    FOSVRCustomPresent(OSVR_ClientContext clientContext) :
+        FRHICustomPresent(nullptr)
+    {
+        mClientContext = osvrClientInit("com.osvr.unreal.plugin.FOSVRCustomPresent");
+    }
+
+    virtual ~FOSVRCustomPresent() {
+        if (mClientContext) {
+            osvrClientShutdown(mClientContext);
+        }
+
+        if (mRenderManager) {
+            osvrDestroyRenderManager(mRenderManager);
+        }
+    }
+
+    // virtual methods from FRHICustomPresent
+
+    virtual void OnBackBufferResize() override {}
+
+    virtual bool Present(int32 &inOutSyncInterval) override {
+        check(IsInRenderingThread());
+        FScopeLock lock(&mOSVRMutex);
+        InitializeImpl();
+        FinishRendering();
+        return true;
+    }
+
+    // implement this in the sub-class
+    virtual bool Initialize() {
+        FScopeLock lock(&mOSVRMutex);
+        return InitializeImpl();
+    }
+
+    virtual bool IsInitialized() {
+        return mInitialized;
+    }
+
+    virtual bool UpdateViewport(const FViewport& InViewport, class FRHIViewport* InViewportRHI) = 0;
+
+    // RenderManager normalizes displays a bit. We create the render target assuming horizontal side-by-side.
+    // RenderManager then rotates that render texture if needed for vertical side-by-side displays.
+    virtual bool CalculateRenderTargetSize(uint32& InOutSizeX, uint32& InOutSizeY) {
+        FScopeLock lock(&mOSVRMutex);
+        return CalculateRenderTargetSizeImpl(InOutSizeX, InOutSizeY);
+    }
+
+    virtual bool AllocateRenderTargetTexture(uint32 index, uint32 sizeX, uint32 sizeY, uint8 format, uint32 numMips, uint32 flags, uint32 targetableTextureFlags, FTexture2DRHIRef& outTargetableTexture, FTexture2DRHIRef& outShaderResourceTexture, uint32 numSamples = 1) = 0;
+
+protected:
+    FCriticalSection mOSVRMutex;
+    std::vector<OSVR_ViewportDescription> mViewportDescriptions;
+    OSVR_RenderParams mRenderParams;
+
+    bool mRenderBuffersNeedToUpdate = true;
+    bool mInitialized = false;
+    OSVR_ClientContext mClientContext = nullptr;
+    OSVR_RenderManager mRenderManager = nullptr;
+
+    virtual bool CalculateRenderTargetSizeImpl(uint32& InOutSizeX, uint32& InOutSizeY) = 0;
+
+    virtual bool InitializeImpl() = 0;
+
+    virtual TGraphicsDevice* GetGraphicsDevice() {
+        auto ret = RHIGetNativeDevice();
+        return reinterpret_cast<TGraphicsDevice*>(ret);
+    }
+
+    virtual void FinishRendering() = 0;
+
+    // abstract methods, implement in DirectX/OpenGL specific subclasses
+    virtual std::string GetGraphicsLibraryName() = 0;
+    virtual bool ShouldFlipY() = 0;
+    virtual void UpdateRenderBuffers() = 0;
+};
+
